# Optimisation: skybox rendering