--- a/src/engine/renderer/glsl_source/skybox_vp.glsl
+++ b/src/engine/renderer/glsl_source/skybox_vp.glsl
@@ -24,7 +24,6 @@ Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 
 IN vec3 		attr_Position;
 
-uniform mat4		u_ModelMatrix;
 uniform mat4		u_ModelViewProjectionMatrix;
 
 OUT(smooth) vec3	var_Position;
@@ -34,7 +33,6 @@ void	main()
 	// transform vertex position into homogenous clip-space
 	gl_Position = u_ModelViewProjectionMatrix * vec4(attr_Position, 1.0);
 
-	// transform position into world space
-	var_Position = (u_ModelMatrix * vec4(attr_Position, 1.0)).xyz;
+	var_Position = attr_Position;
 }