… src/engine/renderer/tr_sky.cpp