Generate a persistent mesh for the skybox consisting of a cube with 8 vertices and 12 triangles.
Old code was creating a mesh consisting of many "blocks" (2 triangles each) on the CPU and uploading it to the GPU *every frame that the sky was rendered*. Not only that, but it performed operations on mapped memory. The mesh itself normally consisted of 1k+ triangles, which in itself is not much for modern hardware, but it's far more than needed and often more than the rest of the view combined. This was horrible for perfomance.

New mesh:
![image](https://github.com/DaemonEngine/Daemon/assets/10687142/57c4887c-e6b0-42da-a30e-a6cf89fc0eb9)
Old mesh example:
![image](https://github.com/DaemonEngine/Daemon/assets/10687142/4730b3af-ba36-4839-b3ac-0ac73158c134)
![image](https://github.com/DaemonEngine/Daemon/assets/10687142/513ac910-9108-44cd-a83c-3ec28e33a29b)

Move cloud generation to fragment shader. The fragment shader will now branch on a uniform to sample different textures, which will likely be faster than switching shader programs. The skybox itself will now always be centered on the view origin. For portals, the last projection matrix obtained when portalLevel == 0 is used because portals use a different projection matrix.

NUKE sky code used for clipping (i. e. generating those sky blocks). I've left the Tess_StageIteratorDepthFill() and Render_DepthFill() there even though they're unused, since I noticed that the engine uses different shader programs for depth pre-pass, so it might be possible to repurpose these functions for a more efficient depth pre-pass.

Fixes #849 and #1021.

Reference points for testing from https://github.com/DaemonEngine/Daemon/issues/849#issuecomment-1925448467:

## classic skybox

map|scene
-|-
plat23|default spectator view

new(FPS(ms))|old(FPS(ms))|r_fastsky 1(FPS(ms))
-|-|-
220(4.54ms)|200(5ms)|200(5ms)

## cloudy skybox

map|scene
-|-
chasm|default spectator view

new(FPS(ms))|old(FPS(ms))|r_fastsky 1(FPS(ms))
-|-|-
160(6.25ms)|150(6.66ms)|150(6.66ms)

## sky portal with classic skybox

map|player viewpos|skybox viewpos
-|-|-
station15|270 37 144 121 0|956 -2450 -3763 121 0 

new(FPS(ms))|old(FPS(ms))|r_fastsky 1(FPS(ms))
-|-|-
180(5.55ms)|180(5.55ms)|200(5ms)

## sky portal with cloudy skybox

map|player viewpos|skybox viewpos
-|-|-
[atcs-2015](https://users.unvanquished.net/~sweet/pkg/map-atcs-2015_0%2Bv3.2-final.dpk)|80 -106 457 50 -2|80 -106 -13362 50 0

new(FPS(ms))|old(FPS(ms))|r_fastsky 1(FPS(ms))
-|-|-
180(5.55ms)|170(5.88ms)|190(5.26ms)

## many sky brushes

map|default spectator view
-|-
testing0|default spectator view

new(FPS(ms))|old(FPS(ms))|r_fastsky 1(FPS(ms))
-|-|-
180(5.55ms)|120(8.33ms)|180(5.55ms)

map|default spectator view
-|-
[dretchstorm2](https://users.unvanquished.net/~sweet/pkg/map-dretchstorm2_2.dpk)|default spectator view

new(FPS(ms))|old(FPS(ms))|r_fastsky 1(FPS(ms))
-|-|-
150(6.66ms)|100(10ms)|150(6.66ms)

Portals:
![unvanquished_2024-02-14_061302_000](https://github.com/DaemonEngine/Daemon/assets/10687142/36e1fd90-b27d-4b83-962f-af788bba1996)
![unvanquished_2024-02-14_061337_000](https://github.com/DaemonEngine/Daemon/assets/10687142/8de8c2ff-1c68-43c4-9a7e-b2bd930bbf44)