@illwieckz 2024-02-14
Also, I love how the  size of the code is now only 33% of the original one… This is mind blowing! 🤯️